sol: proper deallocation of struct sol_config
The SOL block was originally designed to have a single instance,
so the code was not accounting for multiple deallocations.

This patch adds the reference count that is employed in the GK and
GT blocks to deal with the issue.

On top of the multiple deallocations, there is a race condition
during the deallocations, so the errors that are logged vary.
The following is only an example of what can be found:
RING: Cannot free ring, not created with rte_ring_create()

The problem becomes more pronounced with the increase of
NUMA nodes since a SOL instance is allocated to each NUMA node.

While Gatekeeper was not finishing cleanly, this issue was ONLY
creating stranger log entries.